recuperation des jsonArray getInfoZones

+++ app/src/main/java/com/example/mapaide/moodmap/Evaluation.java
@@ -5,15 +5,15 @@ package com.example.mapaide.moodmap;
  */
 public class Evaluation {
     private String critere;
-    private float longitude;
-    private float latitude;
+    private double longitude;
+    private double latitude;
     private int score;
 
     public Evaluation(){
 
     }
 
-    public Evaluation(String c, float log, float lt ,int s){
+    public Evaluation(String c, double log, double lt ,int s){
         this.critere = c;
         this.longitude = log;
         this.latitude = lt;
@@ -33,11 +33,11 @@ public class Evaluation {
         this.score = s;
     }
 
-    public void setLongitude(float lg ) {
+    public void setLongitude(double lg ) {
         this.longitude = lg;
     }
 
-    public void setLatitude(float lat ) {
+    public void setLatitude(double lat ) {
         this.latitude = lat;
     }
 
@@ -49,10 +49,10 @@ public class Evaluation {
         return score;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
-    public float getLatitude(){
+    public double getLatitude(){
         return latitude;
     }
 }
